+++ influxql/select_test.go
@@ -1250,7 +1250,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 		Points    [][]influxql.Point
 	}{
 		{
-			Name:      "rhs binary add",
+			Name:      "rhs binary add number",
+			Statement: `SELECT value + 2.0 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 21}},
+			},
+		},
+		{
+			Name:      "rhs binary add integer",
 			Statement: `SELECT value + 2 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
@@ -1259,7 +1268,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary add",
+			Name:      "lhs binary add number",
+			Statement: `SELECT 2.0 + value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 21}},
+			},
+		},
+		{
+			Name:      "lhs binary add integer",
 			Statement: `SELECT 2 + value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
@@ -1277,7 +1295,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary multiply",
+			Name:      "rhs binary multiply number",
+			Statement: `SELECT value * 2.0 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 38}},
+			},
+		},
+		{
+			Name:      "rhs binary multiply integer",
 			Statement: `SELECT value * 2 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
@@ -1286,7 +1313,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary multiply",
+			Name:      "lhs binary multiply number",
+			Statement: `SELECT 2.0 * value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 38}},
+			},
+		},
+		{
+			Name:      "lhs binary multiply integer",
 			Statement: `SELECT 2 * value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
@@ -1304,7 +1340,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary subtract",
+			Name:      "rhs binary subtract number",
+			Statement: `SELECT value - 2.0 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 18}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 8}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 17}},
+			},
+		},
+		{
+			Name:      "rhs binary subtract integer",
 			Statement: `SELECT value - 2 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 18}},
@@ -1313,7 +1358,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary subtract",
+			Name:      "lhs binary subtract number",
+			Statement: `SELECT 2.0 - value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: -18}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: -8}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: -17}},
+			},
+		},
+		{
+			Name:      "lhs binary subtract integer",
 			Statement: `SELECT 2 - value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: -18}},
@@ -1331,7 +1385,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary division",
+			Name:      "rhs binary division number",
+			Statement: `SELECT value / 2.0 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 10}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 5}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: float64(19) / 2}},
+			},
+		},
+		{
+			Name:      "rhs binary division integer",
 			Statement: `SELECT value / 2 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 10}},
@@ -1340,7 +1403,16 @@ func TestSelect_BinaryExpr_Float(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary division",
+			Name:      "lhs binary division number",
+			Statement: `SELECT 38.0 / value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 1.9}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 3.8}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 2}},
+			},
+		},
+		{
+			Name:      "lhs binary division integer",
 			Statement: `SELECT 38 / value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 1.9}},
@@ -1391,8 +1463,8 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 		Points    [][]influxql.Point
 	}{
 		{
-			Name:      "rhs binary add",
-			Statement: `SELECT value + 2 FROM cpu`,
+			Name:      "rhs binary add number",
+			Statement: `SELECT value + 2.0 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
@@ -1400,12 +1472,30 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary add",
-			Statement: `SELECT 2 + value FROM cpu`,
+			Name:      "rhs binary add integer",
+			Statement: `SELECT value + 2 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: 21}},
+			},
+		},
+		{
+			Name:      "lhs binary add number",
+			Statement: `SELECT 2.0 + value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 21}},
+			},
+		},
+		{
+			Name:      "lhs binary add integer",
+			Statement: `SELECT 2 + value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: 22}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: 12}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: 21}},
 			},
 		},
 		{
@@ -1418,8 +1508,8 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary multiply",
-			Statement: `SELECT value * 2 FROM cpu`,
+			Name:      "rhs binary multiply number",
+			Statement: `SELECT value * 2.0 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
@@ -1427,12 +1517,30 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary multiply",
-			Statement: `SELECT 2 * value FROM cpu`,
+			Name:      "rhs binary multiply integer",
+			Statement: `SELECT value * 2 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: 38}},
+			},
+		},
+		{
+			Name:      "lhs binary multiply number",
+			Statement: `SELECT 2.0 * value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 38}},
+			},
+		},
+		{
+			Name:      "lhs binary multiply integer",
+			Statement: `SELECT 2 * value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: 40}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: 20}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: 38}},
 			},
 		},
 		{
@@ -1445,8 +1553,8 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary subtract",
-			Statement: `SELECT value - 2 FROM cpu`,
+			Name:      "rhs binary subtract number",
+			Statement: `SELECT value - 2.0 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 18}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 8}},
@@ -1454,12 +1562,30 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary subtract",
-			Statement: `SELECT 2 - value FROM cpu`,
+			Name:      "rhs binary subtract integer",
+			Statement: `SELECT value - 2 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: 18}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: 8}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: 17}},
+			},
+		},
+		{
+			Name:      "lhs binary subtract number",
+			Statement: `SELECT 2.0 - value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: -18}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: -8}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: -17}},
+			},
+		},
+		{
+			Name:      "lhs binary subtract integer",
+			Statement: `SELECT 2 - value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.IntegerPoint{Name: "cpu", Time: 0 * Second, Value: -18}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 5 * Second, Value: -8}},
+				{&influxql.IntegerPoint{Name: "cpu", Time: 9 * Second, Value: -17}},
 			},
 		},
 		{
@@ -1472,8 +1598,8 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "rhs binary division",
-			Statement: `SELECT value / 2 FROM cpu`,
+			Name:      "rhs binary division number",
+			Statement: `SELECT value / 2.0 FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 10}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 5}},
@@ -1481,12 +1607,30 @@ func TestSelect_BinaryExpr_Integer(t *testing.T) {
 			},
 		},
 		{
-			Name:      "lhs binary division",
-			Statement: `SELECT 38 / value FROM cpu`,
+			Name:      "rhs binary division integer",
+			Statement: `SELECT value / 2 FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 10}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 5}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: float64(19) / 2}},
+			},
+		},
+		{
+			Name:      "lhs binary division number",
+			Statement: `SELECT 38.0 / value FROM cpu`,
 			Points: [][]influxql.Point{
 				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 1.9}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 3.8}},
 				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 2.0}},
+			},
+		},
+		{
+			Name:      "lhs binary division integer",
+			Statement: `SELECT 38 / value FROM cpu`,
+			Points: [][]influxql.Point{
+				{&influxql.FloatPoint{Name: "cpu", Time: 0 * Second, Value: 1.9}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 5 * Second, Value: 3.8}},
+				{&influxql.FloatPoint{Name: "cpu", Time: 9 * Second, Value: 2}},
 			},
 		},
 		{